[maths] Move utils/omnes-factor to maths and fix include statements.
