+++ eos/maths/Makefile.am
@@ -19,6 +19,7 @@ libeosmaths_la_SOURCES = \
 	matrix.hh \
 	multiplepolylog-li22.cc multiplepolylog-li22.hh \
 	multiplepolylog-li22-const.cc multiplepolylog-li22-const.hh \
+	omnes-factor.hh omnes-factor-impl.hh \
 	outer-function.hh outer-function.cc \
 	polylog.cc polylog.hh \
 	power-of.hh \
@@ -51,6 +52,7 @@ include_eos_utils_HEADERS = \
 	matrix.hh \
 	multiplepolylog-li22.hh \
 	multiplepolylog-li22-const.hh \
+	omnes-factor.hh omnes-factor-impl.hh \
 	outer-function.hh \
 	polylog.hh \
 	power-of.hh \
@@ -69,6 +71,7 @@ TESTS = \
 	legendre-polynomial-vector_TEST \
 	matrix_TEST \
 	multiplepolylog_TEST \
+	omnes-factor_TEST \
 	outer-function_TEST \
 	polylog_TEST \
 	power_of_TEST \
@@ -101,6 +104,8 @@ legendre_polynomial_vector_TEST_SOURCES = legendre-polynomial-vector_TEST.cc
 matrix_TEST_SOURCES = matrix_TEST.cc
 
 multiplepolylog_TEST_SOURCES = multiplepolylog_TEST.cc
+
+omnes_factor_TEST_SOURCES = omnes-factor_TEST.cc
 
 outer_function_TEST_SOURCES = outer-function_TEST.cc
