+++ eos/maths/omnes-factor.hh
